Remove call to close sockets on disconnect
In a live environment I've observed sporadic failures where
the call to open sockets would fail because the port was,
supposedly, already open ({{:error, :eaddrinuse}). I'm not sure
why this occurs as it does seem to call close socket every time,
but I suppose this could be a race condition either in the VM
or at the OS level. Or the socket may simply be allocated
to another application in the time between closing and
opening the socket.

Closing sockets was done to avoid receiving messages meant
for the previous connection after disconnecting and reconnecting.
However, this is not enough to avoid receiving obsolete messages,
because of network latency. Thus logic was added to handle these
cases anyway.

Since logic exists to handle the receipt of messages when
communication_channel_id is nil or does not match that of the
message, I have decided to stop closing the ports, in an attempt
to avoid the issue with opening already open ports.